test(#7): Add tests for the get_fuv function
This adds two unit tests for the get_fuv function. However, there is uncertainty with regards to replicating the example case given in the assignment: either the implementation and my understanding of the functionality of the function are wrong or there is missing info in the example. (See comment in line 41)